// src/types/calculator.ts
export interface CalcForm {
  netProfit: number;
  expenses: number;
  sponsorship: number;
  commissionRate: number;
  totalRegistrations: number;
  priceIncreaseRate: number;
  earlyBirdPct: number;
  standardPct: number;
  fullPricePct: number;
  preConEnabled: boolean;
  preConNetProfit: number;
  honorariumPerAttendee: number;
  preConAttendancePct: number;
}